fix(dataplanes): Moves dataplanes to use state instead of health
