+++ src/app/data-planes/data/MeshGatewayDataplane.ts
@@ -1,0 +1,70 @@
+import type {
+  DataPlaneProxyStatus as DataplaneStatusCount,
+  MeshGatewayDataplane as PartialMeshGatewayDataplane,
+  MeshGatewayListenerEntry,
+  MeshGatewayRouteEntry,
+  Meta,
+} from '@/types/index.d'
+
+export type MeshGatewayDataplane = PartialMeshGatewayDataplane & {
+  listenerEntries: MeshGatewayListenerEntry[]
+  routePolicies: Meta[]
+}
+
+export const MeshGatewayDataplane = {
+  fromObject(partialMeshGatewayDataplane: PartialMeshGatewayDataplane): MeshGatewayDataplane {
+    const listenerEntries = getListenerEntries(partialMeshGatewayDataplane)
+    const routePolicies = Object.values(partialMeshGatewayDataplane.policies ?? {}).map(({ mesh, name, type }) => ({ mesh, name, type }))
+
+    return {
+      ...partialMeshGatewayDataplane,
+      listenerEntries,
+      routePolicies,
+    }
+  },
+}
+
+function getListenerEntries(meshGatewayDataplane: PartialMeshGatewayDataplane): MeshGatewayListenerEntry[] {
+  const meshGatewayListenerEntries: MeshGatewayListenerEntry[] = []
+
+  const listeners = meshGatewayDataplane.listeners ?? []
+  for (const listener of listeners) {
+    for (const host of listener.hosts) {
+      for (const route of host.routes) {
+        const routeEntries: MeshGatewayRouteEntry[] = []
+
+        for (const destination of route.destinations) {
+          const origins = Object.values(destination.policies ?? {}).map(({ mesh, name, type }) => ({ mesh, name, type }))
+
+          routeEntries.push({
+            route: {
+              mesh: meshGatewayDataplane.gateway.mesh,
+              name: route.route,
+              type: 'MeshGatewayRoute',
+            },
+            service: destination.tags['kuma.io/service'],
+            origins,
+          })
+        }
+
+        meshGatewayListenerEntries.push({
+          protocol: listener.protocol,
+          port: listener.port,
+          hostName: host.hostName,
+          routeEntries,
+        })
+      }
+    }
+  }
+
+  return meshGatewayListenerEntries
+}
+
+export function getDataplaneStatusCounts({ total = 0, online = 0, partiallyDegraded = 0, offline = 0 }: DataplaneStatusCount): Required<DataplaneStatusCount> {
+  return {
+    total,
+    online,
+    partiallyDegraded,
+    offline,
+  }
+}
